requirements.txt
# Core dependencies
beautifulsoup4==4.12.3
google-generativeai==0.3.2
google-api-python-client==2.120.0
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
trafilatura==1.6.4
google-auth==2.38.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
soupsieve==2.6
urllib3==2.3.0
solana==0.36.6
solders>=0.23.0,<0.27.0
base58>=2.1.1

# Other dependencies
gTTS>=2.3,<2.4
tenacity>=8.2.3
httpx>=0.26.0
